+++ lektor_tekir/static/color-mode.js
@@ -1,13 +1,12 @@
 window.addEventListener("DOMContentLoaded", () => {
     const colorMode = localStorage.getItem("color-mode");
-    console.log(colorMode);
     if (colorMode) {
         document.documentElement.setAttribute("color-mode", colorMode);
     }
 
-    document.querySelectorAll("button.color-mode").forEach(btn => {
-        btn.addEventListener("click", (e) => {
-            if (e.currentTarget.classList.contains("light")) {
+    document.querySelectorAll("button.color-mode").forEach((el) => {
+        el.addEventListener("click", (ev) => {
+            if (ev.currentTarget.classList.contains("light")) {
                 document.documentElement.setAttribute("color-mode", "light");
                 localStorage.setItem("color-mode", "light");
             } else {