+++ lektor_tekir/static/delete-confirm.js
@@ -1,25 +1,24 @@
-window.addEventListener("DOMContentLoaded", (e) => {
+window.addEventListener("DOMContentLoaded", (ev) => {
     const deleteDialog = document.getElementById("delete-confirm");
 
     document.querySelectorAll(".delete-content").forEach((el) => {
-        el.addEventListener("click", () => {
-            document.getElementById("delete-continue").setAttribute(
-              "hx-vals",
-              "js:{'path': document.getElementById('delete-item-path').innerHTML}"
-            );
+        el.addEventListener("click", (ev) => {
+            document.getElementById("delete-continue").setAttribute("hx-include", "#" + el.closest("form").id);
             deleteDialog.showModal();
         });
     });
 
-    document.getElementById("delete-continue").addEventListener("click", () => {
-        const deletedItemPath = document.getElementById('delete-item-path').innerHTML;
-        const deletedButton = document.querySelector("button[data-path='" + deletedItemPath + "']");
-        const deletedRow = deletedButton.closest("tr");
-        deletedRow.remove();
+    document.getElementById("delete-continue").addEventListener("click", (evt) => {
+        evt.preventDefault();
+        const formId = evt.target.getAttribute("hx-include");
+        document.querySelector(formId).querySelectorAll("input:checked").forEach((el) => {
+            el.closest("tr").remove();
+        });
         deleteDialog.close();
     });
 
-    document.getElementById("delete-cancel").addEventListener("click", () => {
+    document.getElementById("delete-cancel").addEventListener("click", (evt) => {
+        evt.preventDefault();
         deleteDialog.close();
     });
 });