+++ CHANGES.rst
@@ -4,6 +4,7 @@ Changes
 0.2 (unreleased)
 ----------------
 
+- Add support for deleting multiple selected content items.
 - Add dark-light mode toggle.
 
 0.1 (2023-07-04)